Adjust the login address to something that is useful for origins outside of the SpiderOak.com domain, and set it to a private var in the new spideroak singleton instance.
Refactor the spideroak object definition so it's a singleton instance
framed with Crockford's module singleton pattern.

This all seems proper and operational, *except* that it's currently blocked
from successful XmlHTTPRequests when situated outside of SpiderOak.com due
to cross-domain XHR request restrictions, evidenced by the browser warning
(showing in the javascript debugging console in Chrome):

   XMLHttpRequest cannot load https://spideroak.com/browse/login.
   Origin http://myriadicity.net is not allowed by Access-Control-Allow-Origin.

This stackoverflow answer describes the problem pretty closely:

   http://stackoverflow.com/questions/3595515/xmlhttprequest-error-origin-null-is-not-allowed-by-access-control-allow-origin#answer-3744697

I could confirm the problem (and determine whether the code I'm using would
work, in the absence of the problem) if I could run the script from a page
visited within the SpiderOak.com domain.

In any case, I'm going to need to get it working from a local file, and
using a jsonP XmlHTTPRequest seems to be the best prospect for doing that.