Clarify name of visit_share_root().
Include type emblems in ContentNode.toString(), so we can tell what kind of
node the subject of error notices. (The browser debugger is able to
indicate the exact type of object, why the HECK doesn't javascript make
that available? Javascript does have good parts, but really, the bad parts
are abhorrent.)