Preserving this spiffy formulation of a settings manager before changing it to something a little less spiffy, but also a little less prone to mishap.
This approach, using getters and setters, provides persistence for
attribute assignments, as property or array elements, *once* you've
registered the attributes. Problem is everything but persistence works if
you fail to do the registration, leaving too much opportunity for
hard-to-uncover bugs.

I think we can't make this approach better because javascript doesn't seem
to provide the means to trap arbitrary property assignments. So i'll be
switching to more mundane .get()/ .set().