The changePage transition machinery for popups is apparently conflicting with my special changePage provisions.  Clicking links in the popups I'm fabricating on-the-fly is not working right.  Instead, the exact link that was clicked to launch the popup (through my special changePage provisions) is being repeated.
I thought that detecting and avoiding the repeat transits would enable the
proper transits to happen, but not so - I don't know how to reach
them. I've decided to use a different, less brittle approach for popups,
fabricating the menus when creating the links, and using a regular
(data-rel="popup") popup link instead of fabricating on-the-fly.  However,
the transit_manager machinery I created and integrated is worth
remembering, hence this branch.

This checkin is constituted by the central transit_manager.