Incorporate transit_manager in PublicRootShareNode.enlisted_room_menu().
Also implement a workaround for the popup transit problem, using onclick()
to circumvent the changePage machinery altogether.

The big drawback: it depends on exposing spideroak object APIs because the
onclick isn't situated within the spideroak object context.

It partly works, and can be fixed - but I think another approach is much
better - create the popup menus in the DOM during creation of the links to
them, and use regular link actions to get there.