An alternative approach to the enlisted menu popup, with an alternative problem.
The popup menu is reliable in this case - it doesn't seem to get shy.  But
we don't stay on the page for very long, because if we dismiss it twice in
a row without taking an action, it causes a traversal to the dashboard.  (I
think the return mechanism is supposed to be resending the original page,
but it's getting confused, and ticking them back.)

The difference in the approach is that we prepare the menu using an
onclick-associated method, and then let the button click activate the
menu (with data-rel="popup").

(To make the data-rel="popup" work, we seem to need to put the popup on the
page where the button is getting clicked.  That's kinda icky, too.)

I'm preserving this in case I get an idea about how to improve matters, or
to adapt something from it.