We can work around the same-origin restriction using a reverse proxy that Frank Sievetsen established at fx5.de.  By visiting https://www.fx5.de/so_html5/SpiderOak.html, we can use the proxy so that the storage server sees visits as coming from within spideroak.com.  This change makes some adjustments so that the login relay stays with the proxy.
We also need a provision for accounts tended by the second SpiderOak
storage server.  I've asked Frank to set that up, and will adjust things
accordingly.

This convolution is conditioned by being in debugging mode.  Plus, it's
good to scope out what all has to be done to provide for variations on the
server address.  It's a bit more than I expected .. specifically because the
proxy adds some path between the server host and the storage and shares
suffixes .. but still very well contained.