Checkin of a temporary change, for debugging and posterity.
Change the share rooms gotten from the

Using PublicRootShareNode as the parent is causing the share room's name to
be left blank.  (The name is properly set if the share room is also among
the account's original shares, making reproducing the problem a bit more
intricate.)

I need to investigate this and see if there is any correction to be done.
I will then revert to using the combo root as the parent, since the back
button for rooms is going to be resolved according to recent_tab, anyway.