Plug tab tracking into footer navbar layout, so the current tab is highlighted.
Currently, the highlight is rudimentary - make the tab text orange.  We
need to instead toggle the tab background gradient, and maybe the icon
color (but the icon color isn't enough, because some compact views lack the
icon).  That will all be done in css.

Also convert method commentary of involved methods to jsdoc.