Switch completely to 2.2.0.  (We're using elements from a clone of the current repository.)
Track change in new versions to using a 'build' script, renamed from
'debug', (The documentation and ancillary scripts aren't yet properly
updated in the master repository.)

Adjust the codova-X.Y.Z.js script used according to the version.  We also
properly fetch it, now, from the copy of the cordova library we're using,
so we have the corresponding version.

Add a --verbose option, so people can see the output from the build and
emulate scripts.