Use copies of the selected brand and color artifacts, in the prototype / development copy, rather than symbolic links.
The script now checks to confirm that the prototype/dev artifact copies are
identical, and bails if they're not.  It also provides a new option,
'--sync-prototype', to do the sync, and mentions that when it bails.

This is all in preparation to commissioning some changes for PhoneGap
extensions, which will best be managed with some similar copying.